Consolidated Stripes and Gradient into a single Pattern struct
The only difference between these two (and other patterns to be added in the future) was the function used to calculate the color at a point. This consolidation heavily reduces code duplication by creating a struct that holds all of that shared functionality, only needing the aforementioned function.

What about patterns with more than 2 colors? Const generics are the answer to that. Looking forward to stabilization of const generics - this Pattern struct will become amazing.

FnPattern does not take references, even if I want it to. I tried really hard, but wasn't able to solve problems with lifetimes... Points and Colors are cheap to copy, however, so this should be an okay change. If this ends up as the bottleneck of this ray-tracer, I think I've done a pretty amazing job.