More generic Texture trait and Patterns can themselves have Patterns
The Texture trait is now generic - its only method is now color_at(). The transformation stuff can be added back later, if required. Currently I view it as Pattern-specific, not Texture-specific. We'll have to see what changes I'll need to make when adding image textures.

The Pattern struct has also become more generic - it can now use any Texture, instead of requiring a Color. This will make it harder to expand the struct with const generics in the future, since each Pattern will have a different size depending on which Texture it is used with... I'll deal with that problem when it arises.